Addcalculator function
