--- src/logic/operator.js
@@ -1,14 +1,0 @@
-import operate from './operate';
-
-function isNumber(item) {
-  return !!item.match(/[0-9]+/);
-}
-
-export default function calculate(obj, buttonName) {
-  if (buttonName === 'AC') {
-    return {
-      total: null,
-      next: null,
-      operation: null,
-    };
-  }
